+++ changelog.md
@@ -2,6 +2,7 @@
 
 #### v1.6.2
 
+* `[fsutil]` Improved working with paths
 * `[fsutil]` Added method ProperPath to windows stub
 
 #### v1.6.1